requirements.txt
streamlit==1.29.0
mysql-connector-python==8.2.0
pandas==2.1.3
path==16.9.0
